refactor(storage): consistent folder structure
