+++ packages/core/manifest/src/storage/services/storage.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common'
 import { kebabize } from '@repo/helpers'
-import { DEFAULT_IMAGE_SIZES, STORAGE_PATH } from '../../../constants'
+import { DEFAULT_IMAGE_SIZES, STORAGE_PATH } from '../../constants'
 
 import * as fs from 'fs'
 import * as mkdirp from 'mkdirp'
