+++ packages/core/manifest/src/upload/upload.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common'
 import { UploadController } from './controllers/upload.controller'
-import { StorageService } from '../storage/services/storage/storage.service'
 import { ManifestModule } from '../manifest/manifest.module'
 import { StorageModule } from '../storage/storage.module'
 import { UploadService } from './services/upload.service'
@@ -8,6 +7,7 @@ import { UploadService } from './services/upload.service'
 @Module({
   imports: [ManifestModule, StorageModule],
   controllers: [UploadController],
-  providers: [StorageService, UploadService]
+  providers: [UploadService],
+  exports: [UploadService]
 })
 export class UploadModule {}
